requirements.txt
flask
flask-cors
numpy
pandas
scikit-learn
joblib
python-dotenv
gunicorn